Created a new folder to work in.  Starting with a full copy of the existing shiny app from the Wolf et al. paper, and a copy of gen.Rda from the Vasilev et al. paper.
Will try to adapt the existing shiny app to work with the new data (gen.Rda) instead of the original Wolf et al. data file (df.Rda).  Later will need to create a new data file that has ALL the data; currently the data for comprehension, speed, and proofreading are in separate data files.  Once they are in a single file we can then use "task" and "measure" to select the right subsets of data to reproduce the original Vasilev et al. analyses.

A couple of possible differences worth noting between the Vasilev vs. Wolf analyses:

1) Wolf et al. aggregate all measures within each study (each paper) before doing the meta-analysis.  Vasilev et al. appear to treat each experiment in a paper as a separate entry in the meta-analysis, not aggregating results within each study.

2) Vasilev et al. appear to use the default prior for mu, which I think is a uniform distribution.  That is not one of the options in the Wolf et al. shiny app.  (And it is an improper prior, which means the bayesian sensitivity analysis would not run.)

3) V. use an improper uniform prior for the variance, which also would prevent the bayesian sensitivity analysis from running.

4) Both appear to use g as the effect size, but V. labels it "g" and variance of g, while W. labels it "yi" and "vi" in their data file.