Version 0.3
Users can add individual studies now, but: (1) nneed to fix the paper number issue for added studies (2) Have only tested it by inputting g and g_var (3) lots of "debugging" lines need to be removed (4) a panel needs to be added to display the current data file (5) a panel needs to be added (or repurposed) to download the current data file (6) Need to change how the "r" value is done.  Use the default estimate if blank, but do not make it the default value in the input form, so that it does not end up in the data file appearing to be actual data from a study.