Warnings (from shinyalert) are now working for tabs that require recalculating bma() or require calculating for new BF robustness plots.
Unfortunately there is a lot of redundancy involved, because the modal (shinyalert) has to set a reactive value in a separate reactive block, which is then referenced in the reactive block that creates the tab output.  Shinyalert does not halt the flow of the code within a block, thus the need for the separate reactive block and reactive value or reactive variable.